+++ .storybook/Layout.js
@@ -4,5 +4,3 @@ import "../styles/app.scss";
 const Layout = ({ children }) => {
   return <React.Fragment>{children}</React.Fragment>;
 };
-
-export default Layout;
