+++ components/forms/Alert/Alert.tsx
@@ -42,5 +42,3 @@ export const Alert = ({
     </div>
   );
 };
-
-export default Alert;
