+++ components/forms/Button/Button.tsx
@@ -53,5 +53,3 @@ export const Button = (
     </button>
   );
 };
-
-export default Button;
