+++ components/forms/Button/DeleteButton.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Button } from "./Button";
+import { Button } from "@components/forms";
 import { useRouter } from "next/router";
 import { useTranslation } from "react-i18next";
 import { logMessage } from "@lib/logger";
@@ -64,5 +64,3 @@ export const DeleteButton = (props: DeleteButtonProps): React.ReactElement => {
     </>
   );
 };
-
-export default DeleteButton;
