+++ components/forms/Checkbox/Checkbox.tsx
@@ -29,5 +29,3 @@ export const Checkbox = (
     </div>
   );
 };
-
-export default Checkbox;
