+++ components/forms/Description/Description.tsx
@@ -25,5 +25,3 @@ export const Description = (props: DescriptionProps): React.ReactElement => {
     </div>
   );
 };
-
-export default Description;
