+++ components/forms/DynamicRow/DynamicRow.tsx
@@ -3,8 +3,7 @@ import classnames from "classnames";
 import { useField } from "formik";
 import { GenerateElement } from "@lib/formBuilder";
 import { FormElement } from "@lib/types";
-import { Button } from "../index";
-import { Description } from "../index";
+import { Button, Description } from "@components/forms";
 import { TFunction } from "next-i18next";
 
 interface DynamicGroupProps {
@@ -161,5 +160,3 @@ export const DynamicGroup = (props: DynamicGroupProps): React.ReactElement => {
     </div>
   );
 };
-
-export default DynamicGroup;
