+++ components/forms/Radio/Radio.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { cleanup, render, screen } from "@testing-library/react";
-import { GenerateElement } from "../../../lib/formBuilder";
+import { GenerateElement } from "@lib/formBuilder";
 import { Formik } from "formik";
 
 const radioButtonData = {
