+++ components/forms/Radio/Radio.tsx
@@ -31,5 +31,3 @@ export const Radio = (
     </div>
   );
 };
-
-export default Radio;
