+++ components/forms/TextArea/TextArea.tsx
@@ -84,5 +84,3 @@ export const TextArea = (
     </>
   );
 };
-
-export default TextArea;
