+++ components/forms/TextInput/TextInput.tsx
@@ -83,5 +83,3 @@ export const TextInput = (
     </>
   );
 };
-
-export default TextInput;
