+++ components/forms/index.ts
@@ -1,6 +1,7 @@
 /* istanbul ignore file */
 export { Alert } from "./Alert/Alert";
 export { ErrorMessage } from "./ErrorMessage/ErrorMessage";
+export { ErrorListItem } from "./ErrorListItem/ErrorListItem";
 export { Button } from "./Button/Button";
 export { DeleteButton } from "./Button/DeleteButton";
 export { Checkbox } from "./Checkbox/Checkbox";
